requirements.txt
anyio==3.6.2
asttokens==2.2.1
autopep8==2.0.0
backcall==0.2.0
bcrypt==4.0.1
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
executing==1.2.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.2
httptools==0.5.0
httpx==0.23.1
idna==3.4
ipdb==0.13.9
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
orjson==3.8.3
parso==0.8.3
passlib==1.7.4
pathspec==0.10.3
pexpect==4.8.0
phonenumbers==8.13.2
pickleshare==0.7.5
platformdirs==2.6.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a29
sqlmodel==0.0.8
stack-data==0.6.2
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
traitlets==5.6.0
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
